requirements.txt
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
get==1.0.3
httplib2==0.9.2
idna==2.6
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
post==1.0.2
public==1.0.3
PyMySQL==0.8.1
PyYAML==3.12
query-string==1.0.2
request==1.0.2
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
